requirements.txt
psycopg2-binary
mysql-connector-python
python-dotenv